Base class for Renderers.
