+++ packages/documentator/src/Markdown/Nodes/Reference/Renderer.php
@@ -2,33 +2,21 @@
 
 namespace LastDragon_ru\LaraASP\Documentator\Markdown\Nodes\Reference;
 
-use LastDragon_ru\LaraASP\Documentator\Package;
+use LastDragon_ru\LaraASP\Documentator\Markdown\XmlRenderer;
 use League\CommonMark\Node\Node;
-use League\CommonMark\Renderer\ChildNodeRendererInterface;
-use League\CommonMark\Renderer\NodeRendererInterface;
-use League\CommonMark\Xml\XmlNodeRendererInterface;
 use Override;
 
 use function assert;
-use function implode;
-use function preg_replace;
 
 /**
  * @internal
  */
-class Renderer implements NodeRendererInterface, XmlNodeRendererInterface {
-    #[Override]
-    public function render(Node $node, ChildNodeRendererInterface $childRenderer): string {
-        assert($node instanceof Block);
-
-        return '';
-    }
-
+class Renderer extends XmlRenderer {
     #[Override]
     public function getXmlTagName(Node $node): string {
         assert($node instanceof Block);
 
-        return Package::Name.'-reference';
+        return 'reference';
     }
 
     /**
@@ -47,19 +35,5 @@ class Renderer implements NodeRendererInterface, XmlNodeRendererInterface {
             'offset'      => $node->getOffset(),
             'location'    => $this->location($node),
         ];
-    }
-
-    private function escape(string $string): string {
-        return preg_replace('/\R/u', '\\n', $string) ?? $string;
-    }
-
-    private function location(Block $node): string {
-        $location = [];
-
-        foreach ($node->getLocation() as $line) {
-            $location[] = '{'.implode(',', [$line->number, $line->offset, $line->length ?? 'null']).'}';
-        }
-
-        return '['.implode(',', $location).']';
     }
 }
